Add files via upload
# Message-Encryption-and-Decryption-in-Python
Message Encryption and Decryption in Python  
In this Repository you will be able to Encrypt or Decrypt any message using Pyhton programming language. 
Rules for Encrypting and Decrypting are as follows -
Rule 1: Enter whether you want to "code or  decode or quit" . 
Rule 2: If you select "code", you will enter any sequence of characters or a string to be encoded.
You will get your message encoded by entering key for encrypting as either of "3 OR 4 OR 6".
Rule 3: For decrypting the same message after encryption, copy the encrypted message and Choose to decode.
After copy pasting the message select key for decrypting( Note: It must be the same key "3 or 4 or 6" that you used for encryption.
You will then receive your decoded message.